scraper/requirements.txt
beautifulsoup4>=4.12.0
feedparser>=6.0.0
pyyaml>=6.0.0
requests>=2.32.0
python-dotenv>=1.0.0
supabase>=2.3.0
lxml>=5.0.0
python-dateutil>=2.8.0
